# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/payment-methods": {
      "post": {
        "operationId": "PaymentMethodsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentMethodDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The payment method has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentMethod"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data provided."
          }
        },
        "summary": "Create a new payment method",
        "tags": [
          "payment-methods"
        ]
      },
      "get": {
        "operationId": "PaymentMethodsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns all payment methods",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentMethod"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all payment methods",
        "tags": [
          "payment-methods"
        ]
      }
    },
    "/api/payment-methods/{id}": {
      "get": {
        "operationId": "PaymentMethodsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Payment method ID (número)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the payment method",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentMethod"
                }
              }
            }
          },
          "404": {
            "description": "Payment method not found"
          }
        },
        "summary": "Get a payment method by ID",
        "tags": [
          "payment-methods"
        ]
      },
      "patch": {
        "operationId": "PaymentMethodsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Payment method ID (número)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentMethodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The payment method has been successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentMethod"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data provided"
          },
          "404": {
            "description": "Payment method not found"
          }
        },
        "summary": "Update a payment method",
        "tags": [
          "payment-methods"
        ]
      },
      "delete": {
        "operationId": "PaymentMethodsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Payment method ID (número)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The payment method has been successfully deleted or inactivated"
          },
          "404": {
            "description": "Payment method not found"
          }
        },
        "summary": "Delete a payment method or mark as inactive if in use",
        "tags": [
          "payment-methods"
        ]
      }
    },
    "/api/payment-terms": {
      "post": {
        "operationId": "PaymentTermsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentTermDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "PaymentTerms"
        ]
      },
      "get": {
        "operationId": "PaymentTermsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PaymentTerms"
        ]
      }
    },
    "/api/payment-terms/{id}": {
      "get": {
        "operationId": "PaymentTermsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PaymentTerms"
        ]
      },
      "patch": {
        "operationId": "PaymentTermsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentTermDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PaymentTerms"
        ]
      },
      "delete": {
        "operationId": "PaymentTermsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "PaymentTerms"
        ]
      }
    },
    "/api/suppliers": {
      "post": {
        "operationId": "SuppliersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupplierDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Fornecedor criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Supplier"
                }
              }
            }
          },
          "400": {
            "description": "Dados de entrada inválidos"
          },
          "409": {
            "description": "Fornecedor com mesmo CNPJ/CPF já existe"
          }
        },
        "summary": "Criar novo fornecedor",
        "tags": [
          "Suppliers"
        ]
      },
      "get": {
        "operationId": "SuppliersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de fornecedores retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Supplier"
                  }
                }
              }
            }
          }
        },
        "summary": "Listar todos os fornecedores",
        "tags": [
          "Suppliers"
        ]
      }
    },
    "/api/suppliers/{id}": {
      "get": {
        "operationId": "SuppliersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do fornecedor",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fornecedor encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Supplier"
                }
              }
            }
          },
          "404": {
            "description": "Fornecedor não encontrado"
          }
        },
        "summary": "Obter fornecedor por ID",
        "tags": [
          "Suppliers"
        ]
      },
      "patch": {
        "operationId": "SuppliersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do fornecedor",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSupplierDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Fornecedor atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Supplier"
                }
              }
            }
          },
          "400": {
            "description": "Dados de entrada inválidos"
          },
          "404": {
            "description": "Fornecedor não encontrado"
          }
        },
        "summary": "Atualizar fornecedor por ID",
        "tags": [
          "Suppliers"
        ]
      },
      "delete": {
        "operationId": "SuppliersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do fornecedor",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Fornecedor removido com sucesso"
          },
          "404": {
            "description": "Fornecedor não encontrado"
          }
        },
        "summary": "Remover fornecedor por ID",
        "tags": [
          "Suppliers"
        ]
      }
    },
    "/api/suppliers/by-document/{cnpjCpf}": {
      "get": {
        "operationId": "SuppliersController_findByDocument",
        "parameters": [
          {
            "name": "cnpjCpf",
            "required": true,
            "in": "path",
            "description": "CNPJ/CPF do fornecedor",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fornecedor encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Supplier"
                }
              }
            }
          },
          "404": {
            "description": "Fornecedor não encontrado"
          }
        },
        "summary": "Obter fornecedor por CNPJ/CPF",
        "tags": [
          "Suppliers"
        ]
      }
    }
  },
  "info": {
    "title": "Serveon API",
    "description": "API RESTful para o sistema Serveon",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePaymentMethodDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "Description of the payment method",
            "example": "Credit Card"
          },
          "code": {
            "type": "string",
            "description": "Code of the payment method",
            "example": "CC"
          },
          "type": {
            "type": "string",
            "description": "Type of payment method",
            "example": "card"
          },
          "active": {
            "type": "boolean",
            "description": "Status of the payment method (active/inactive)",
            "example": true,
            "default": true
          }
        },
        "required": [
          "description"
        ]
      },
      "PaymentMethod": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique ID of the payment method",
            "example": 1
          },
          "description": {
            "type": "string",
            "description": "Description of the payment method",
            "example": "Credit Card"
          },
          "code": {
            "type": "string",
            "description": "Code of the payment method",
            "example": "CC"
          },
          "type": {
            "type": "string",
            "description": "Type of payment method",
            "example": "card"
          },
          "active": {
            "type": "boolean",
            "description": "Status of the payment method (active/inactive)",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation date",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update date",
            "example": "2023-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "description",
          "active",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdatePaymentMethodDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "Description of the payment method",
            "example": "Credit Card"
          },
          "code": {
            "type": "string",
            "description": "Code of the payment method",
            "example": "CC"
          },
          "type": {
            "type": "string",
            "description": "Type of payment method",
            "example": "card"
          },
          "active": {
            "type": "boolean",
            "description": "Status of the payment method (active/inactive)",
            "example": true,
            "default": true
          }
        }
      },
      "CreatePaymentTermDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePaymentTermDto": {
        "type": "object",
        "properties": {}
      },
      "CreateSupplierDto": {
        "type": "object",
        "properties": {
          "cnpjCpf": {
            "type": "string",
            "description": "CNPJ, CPF ou número de documento internacional do fornecedor",
            "example": "12345678901234"
          },
          "tipo": {
            "type": "string",
            "description": "Tipo de fornecedor: F para Pessoa Física, J para Pessoa Jurídica",
            "example": "J",
            "enum": [
              "F",
              "J"
            ]
          },
          "isEstrangeiro": {
            "type": "boolean",
            "description": "Indica se é um fornecedor estrangeiro",
            "example": false,
            "default": false
          },
          "tipoDocumento": {
            "type": "string",
            "description": "Tipo de documento para fornecedores estrangeiros",
            "example": "passport",
            "enum": [
              "passport",
              "tax_id",
              "national_id",
              "other"
            ]
          },
          "razaoSocial": {
            "type": "string",
            "description": "Razão Social ou Nome Completo",
            "example": "Empresa XYZ Ltda"
          },
          "nomeFantasia": {
            "type": "string",
            "description": "Nome Fantasia",
            "example": "XYZ Tecnologia"
          },
          "inscricaoEstadual": {
            "type": "string",
            "description": "Inscrição Estadual ou informações complementares do documento",
            "example": "123456789"
          },
          "inscricaoMunicipal": {
            "type": "string",
            "description": "Inscrição Municipal ou registro comercial internacional",
            "example": "123456"
          },
          "paisId": {
            "type": "number",
            "description": "ID do país",
            "example": 1
          },
          "estadoId": {
            "type": "number",
            "description": "ID do estado/província",
            "example": 1
          },
          "endereco": {
            "type": "string",
            "description": "Endereço",
            "example": "Av. Brasil"
          },
          "numero": {
            "type": "string",
            "description": "Número do endereço",
            "example": "123"
          },
          "complemento": {
            "type": "string",
            "description": "Complemento do endereço",
            "example": "Sala 101"
          },
          "bairro": {
            "type": "string",
            "description": "Bairro",
            "example": "Centro"
          },
          "cidadeId": {
            "type": "number",
            "description": "ID da cidade",
            "example": 1
          },
          "cep": {
            "type": "string",
            "description": "CEP",
            "example": "12345-678"
          },
          "telefone": {
            "type": "string",
            "description": "Telefone",
            "example": "(11) 1234-5678"
          },
          "email": {
            "type": "string",
            "description": "Email",
            "example": "contato@empresa.com"
          },
          "condicaoPagamentoId": {
            "type": "number",
            "description": "ID da condição de pagamento",
            "example": 1
          },
          "website": {
            "type": "string",
            "description": "Website do fornecedor",
            "example": "https://www.empresa.com"
          },
          "observacoes": {
            "type": "string",
            "description": "Observações sobre o fornecedor",
            "example": "Fornecedor preferencial para materiais de escritório"
          },
          "responsavel": {
            "type": "string",
            "description": "Nome do responsável pelo contato",
            "example": "João Silva"
          },
          "celularResponsavel": {
            "type": "string",
            "description": "Celular do responsável",
            "example": "(11) 98765-4321"
          }
        },
        "required": [
          "cnpjCpf",
          "tipo",
          "isEstrangeiro",
          "razaoSocial"
        ]
      },
      "Supplier": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "ID único do fornecedor",
            "example": 1
          },
          "cnpjCpf": {
            "type": "string",
            "description": "CNPJ, CPF ou documento internacional do fornecedor",
            "example": "12345678901234"
          },
          "tipo": {
            "type": "string",
            "description": "Tipo de fornecedor: F para Pessoa Física, J para Pessoa Jurídica",
            "example": "J",
            "enum": [
              "F",
              "J"
            ]
          },
          "isEstrangeiro": {
            "type": "boolean",
            "description": "Indica se é um fornecedor estrangeiro",
            "example": false
          },
          "tipoDocumento": {
            "type": "string",
            "description": "Tipo de documento para fornecedores estrangeiros",
            "example": "passport"
          },
          "razaoSocial": {
            "type": "string",
            "description": "Razão Social ou Nome Completo",
            "example": "Empresa XYZ Ltda"
          },
          "nomeFantasia": {
            "type": "string",
            "description": "Nome Fantasia",
            "example": "XYZ Tecnologia"
          },
          "inscricaoEstadual": {
            "type": "string",
            "description": "Inscrição Estadual ou informações complementares do documento",
            "example": "123456789"
          },
          "inscricaoMunicipal": {
            "type": "string",
            "description": "Inscrição Municipal ou registro comercial internacional",
            "example": "123456"
          },
          "paisId": {
            "type": "number",
            "description": "ID do país",
            "example": 1
          },
          "paisNome": {
            "type": "string",
            "description": "Nome do país",
            "example": "Brasil"
          },
          "estadoId": {
            "type": "number",
            "description": "ID do estado/província",
            "example": 1
          },
          "endereco": {
            "type": "string",
            "description": "Endereço",
            "example": "Av. Brasil"
          },
          "numero": {
            "type": "string",
            "description": "Número do endereço",
            "example": "123"
          },
          "complemento": {
            "type": "string",
            "description": "Complemento do endereço",
            "example": "Sala 101"
          },
          "bairro": {
            "type": "string",
            "description": "Bairro",
            "example": "Centro"
          },
          "cidadeId": {
            "type": "number",
            "description": "ID da cidade",
            "example": 1
          },
          "cidadeNome": {
            "type": "string",
            "description": "Nome da cidade",
            "example": "São Paulo"
          },
          "uf": {
            "type": "string",
            "description": "UF da cidade",
            "example": "SP"
          },
          "cep": {
            "type": "string",
            "description": "CEP",
            "example": "12345-678"
          },
          "telefone": {
            "type": "string",
            "description": "Telefone",
            "example": "(11) 1234-5678"
          },
          "email": {
            "type": "string",
            "description": "Email",
            "example": "contato@empresa.com"
          },
          "ativo": {
            "type": "boolean",
            "description": "Status do fornecedor (ativo/inativo)",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "description": "Data de criação",
            "example": "2023-01-01T00:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "Data da última atualização",
            "example": "2023-01-01T00:00:00Z"
          },
          "condicaoPagamentoId": {
            "type": "number",
            "description": "ID da condição de pagamento",
            "example": 1
          },
          "condicaoPagamentoNome": {
            "type": "string",
            "description": "Nome da condição de pagamento",
            "example": "À Vista"
          },
          "website": {
            "type": "string",
            "description": "Website do fornecedor",
            "example": "https://www.empresa.com"
          },
          "observacoes": {
            "type": "string",
            "description": "Observações sobre o fornecedor",
            "example": "Fornecedor preferencial para materiais de escritório"
          },
          "responsavel": {
            "type": "string",
            "description": "Nome do responsável pelo contato",
            "example": "João Silva"
          },
          "celularResponsavel": {
            "type": "string",
            "description": "Celular do responsável",
            "example": "(11) 98765-4321"
          }
        },
        "required": [
          "id",
          "cnpjCpf",
          "tipo",
          "isEstrangeiro",
          "razaoSocial",
          "nomeFantasia",
          "inscricaoEstadual",
          "inscricaoMunicipal",
          "paisId",
          "paisNome",
          "estadoId",
          "endereco",
          "numero",
          "complemento",
          "bairro",
          "cidadeId",
          "cidadeNome",
          "uf",
          "cep",
          "telefone",
          "email",
          "ativo",
          "createdAt",
          "updatedAt",
          "condicaoPagamentoId",
          "condicaoPagamentoNome",
          "website",
          "observacoes",
          "responsavel",
          "celularResponsavel"
        ]
      },
      "UpdateSupplierDto": {
        "type": "object",
        "properties": {
          "ativo": {
            "type": "boolean",
            "description": "Status ativo/inativo do fornecedor",
            "example": true
          }
        }
      }
    }
  }
}